fix(titanic): looked into potential improvements
I was having trouble improving on my initial score of 0.779, so I wanted to see what else I could do to improve either the feature engineering or the model hyper parameters for better performance. My goal is to at least get 0.815, which seems to be the highest you can get without cheating.